refactor: modify constants value file name
