+++ src/utils/randoms.js
@@ -1,4 +1,4 @@
-import { MAX_RANDOM_NUMBER } from '../racingcar/constatns';
+import { MAX_RANDOM_NUMBER } from '../racingcar/constatns/values';
 
 const randomNumber = () => Math.floor(Math.random() * MAX_RANDOM_NUMBER + 1);
